make separate traffic_light_compliance_checker library
- move tl compliance check code from trajecroy_validator filters to a separate package autoware_traffic_light_compliance_checker
- update trajectory_validator dependencies and remove obsolete filter files
