+++ common/autoware_traffic_light_compliance_checker/README.md
@@ -1,0 +1,3 @@
+# Traffic Light Compliance Checker
+
+Library for checking whether a planned trajectory complies with traffic light signals along the route.
