+++ common/autoware_traffic_light_compliance_checker/include/autoware/traffic_light_compliance_checker/traffic_light_compliance_checker.hpp
@@ -12,82 +12,26 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef AUTOWARE__TRAJECTORY_VALIDATOR__FILTERS__TRAFFIC_RULE__TRAFFIC_LIGHT_COMPLIANCE_CHECKER_HPP_
-#define AUTOWARE__TRAJECTORY_VALIDATOR__FILTERS__TRAFFIC_RULE__TRAFFIC_LIGHT_COMPLIANCE_CHECKER_HPP_
+#ifndef AUTOWARE__TRAFFIC_LIGHT_COMPLIANCE_CHECKER__TRAFFIC_LIGHT_COMPLIANCE_CHECKER_HPP_
+#define AUTOWARE__TRAFFIC_LIGHT_COMPLIANCE_CHECKER__TRAFFIC_LIGHT_COMPLIANCE_CHECKER_HPP_
+
+#include "autoware/traffic_light_compliance_checker/structs.hpp"
 
 #include <autoware/vehicle_info_utils/vehicle_info.hpp>
 #include <tl_expected/expected.hpp>
 
 #include <autoware_perception_msgs/msg/traffic_light_group_array.hpp>
 #include <autoware_planning_msgs/msg/lanelet_route.hpp>
-#include <autoware_planning_msgs/msg/trajectory_point.hpp>
 
 #include <lanelet2_core/LaneletMap.h>
-#include <lanelet2_core/geometry/LineString.h>
 
 #include <optional>
 #include <string>
 #include <utility>
 #include <vector>
 
-namespace autoware::trajectory_validator::traffic_light_filter
+namespace autoware::traffic_light_compliance_checker
 {
-
-/// @brief input data for traffic light compliance check
-struct Inputs
-{
-  std::vector<autoware_planning_msgs::msg::TrajectoryPoint> trajectory;
-  lanelet::LaneletMapPtr map;
-  autoware_planning_msgs::msg::LaneletRoute route;
-  autoware_perception_msgs::msg::TrafficLightGroupArray signals;
-  double current_velocity;
-  double current_acceleration;
-  std::vector<int64_t> force_reject_amber_ids;
-};
-
-/// @brief information about a stop line and its associated traffic light
-struct StopLineInfo
-{
-  lanelet::BasicLineString2d line;
-  int64_t traffic_light_id;
-};
-
-/// @brief type of traffic light violation
-enum class ViolationType { RED_LIGHT, AMBER_LIGHT };
-
-/// @brief violation detail
-struct Violation
-{
-  ViolationType type;
-  lanelet::BasicLineString2d stop_line;
-  int64_t traffic_light_id;
-};
-
-/// @brief result of compliance check
-struct ComplianceResult
-{
-  std::vector<Violation> violations;
-};
-
-/// @brief parameters for traffic light compliance check
-struct Parameters
-{
-  double deceleration_limit;
-  double jerk_limit;
-  double delay_response_time;
-  double crossing_time_limit;
-  bool treat_amber_light_as_red_light;
-  double stop_overshoot_margin;
-  double stable_duration_threshold_red;
-  double stable_duration_threshold_amber;
-  double amber_rejection_hysteresis_duration;
-  double ego_stopped_velocity_threshold;
-  struct CheckedTrajectoryLength
-  {
-    double deceleration_limit;
-    double jerk_limit;
-  } checked_trajectory_length;
-};
 
 /// @brief class to check if a trajectory complies with traffic lights
 class TrafficLightComplianceChecker
@@ -135,6 +79,6 @@ private:
   vehicle_info_utils::VehicleInfo vehicle_info_;
 };
 
-}  // namespace autoware::trajectory_validator::traffic_light_filter
-// NOLINTNEXTLINE
-#endif  // AUTOWARE__TRAJECTORY_VALIDATOR__FILTERS__TRAFFIC_RULE__TRAFFIC_LIGHT_COMPLIANCE_CHECKER_HPP_
+}  // namespace autoware::traffic_light_compliance_checker
+
+#endif  // AUTOWARE__TRAFFIC_LIGHT_COMPLIANCE_CHECKER__TRAFFIC_LIGHT_COMPLIANCE_CHECKER_HPP_
