+++ planning/autoware_trajectory_modifier/launch/trajectory_modifier.launch.xml
@@ -7,17 +7,17 @@
   <arg name="input_pointcloud" default="/perception/obstacle_segmentation/pointcloud"/>
   <arg name="output_candidate_trajectories" default="~/output/candidate_trajectories"/>
 
-  <node_container pkg="rclcpp_components" exec="component_container" name="trajectory_modifier_container" namespace="">
-    <composable_node pkg="autoware_glog_component" plugin="autoware::glog_component::GlogComponent" name="glog_component" namespace=""/>
-    <composable_node pkg="autoware_trajectory_modifier" plugin="autoware::trajectory_modifier::TrajectoryModifier" name="trajectory_modifier" namespace="">
-      <param from="$(var trajectory_modifier_param_path)"/>
-      <remap from="~/input/objects" to="$(var input_objects)"/>
-      <remap from="~/input/pointcloud" to="$(var input_pointcloud)"/>
-      <remap from="~/input/candidate_trajectories" to="$(var input_candidate_trajectories)"/>
-      <remap from="~/input/odometry" to="$(var input_odometry)"/>
-      <remap from="~/input/acceleration" to="$(var input_acceleration)"/>
-      <remap from="~/output/candidate_trajectories" to="$(var output_candidate_trajectories)"/>
-      <extra_arg name="use_intra_process_comms" value="false"/>
-    </composable_node>
-  </node_container>
+  <!-- <node_container pkg="rclcpp_components" exec="component_container" name="trajectory_modifier_container" namespace="">
+    <composable_node pkg="autoware_glog_component" plugin="autoware::glog_component::GlogComponent" name="glog_component" namespace=""/> -->
+  <node pkg="autoware_trajectory_modifier" exec="autoware_trajectory_modifier_node" name="trajectory_modifier" namespace="" launch-prefix="konsole -e gdb -ex run --args">
+    <param from="$(var trajectory_modifier_param_path)"/>
+    <remap from="~/input/objects" to="$(var input_objects)"/>
+    <remap from="~/input/pointcloud" to="$(var input_pointcloud)"/>
+    <remap from="~/input/candidate_trajectories" to="$(var input_candidate_trajectories)"/>
+    <remap from="~/input/odometry" to="$(var input_odometry)"/>
+    <remap from="~/input/acceleration" to="$(var input_acceleration)"/>
+    <remap from="~/output/candidate_trajectories" to="$(var output_candidate_trajectories)"/>
+    <!-- <extra_arg name="use_intra_process_comms" value="false"/> -->
+  </node>
+  <!-- </node_container> -->
 </launch>
