+++ planning/autoware_trajectory_validator/CMakeLists.txt
@@ -43,7 +43,6 @@ ament_auto_add_library(autoware_trajectory_validator_plugins SHARED
   src/filters/safety/collision_check_filter/assessment.cpp
   src/filters/safety/collision_check_filter/trajectory_utils.cpp
   src/filters/traffic_rule/traffic_light_filter.cpp
-  src/filters/traffic_rule/traffic_light_compliance_checker.cpp
 )
 
 set_target_properties(autoware_trajectory_validator_plugins PROPERTIES CXX_VISIBILITY_PRESET default)
