+++ planning/autoware_trajectory_validator/package.xml
@@ -33,6 +33,7 @@
   <depend>autoware_perception_msgs</depend>
   <depend>autoware_planning_factor_interface</depend>
   <depend>autoware_planning_msgs</depend>
+  <depend>autoware_traffic_light_compliance_checker</depend>
   <depend>autoware_traffic_light_utils</depend>
   <depend>autoware_utils_debug</depend>
   <depend>autoware_utils_diagnostics</depend>
